refactor: move `Subclaim` and inline `Subclaim::create`
