+++ crates/proof-of-sql/src/proof_primitive/sumcheck/mod.rs
@@ -6,9 +6,6 @@ pub use proof::SumcheckProof;
 mod prover_state;
 use prover_state::ProverState;
 
-mod subclaim;
-pub use subclaim::Subclaim;
-
 mod prover_round;
 use prover_round::prove_round;
